// src/items/curios.gen.ts
export type CurioBrew = 'heal' | 'mana' | 'elixir' | 'haste' | 'might' | 'stone';
// ...
export const GEN_ORES: ReadonlyArray<readonly [string, number]> = [
  ['o01', 1],
  ['o02', 2],
  ['o03', 3],
  ['o04', 4],
  ['o05', 5],
  ['o06', 6],
  ['o07', 7],
  ['o08', 8],
  ['o09', 9],
  ['o12', 12],
  ['o13', 13],
  ['o14', 14],
  ['o15', 15],
  ['o16', 16],
  ['o17', 17],
  ['o18', 18],
  ['o19', 19],
  ['o20', 20],
  ['o22', 22],
  ['o23', 23],
  ['o24', 24],
  ['o25', 25],
  ['o26', 26],
  ['o27', 27],
  ['o29', 29],
  ['o30', 30],
  ['o31', 31],
  ['o32', 32],
  ['o33', 33],
  ['o34', 34],
  ['o35', 35],
  ['o36', 36],
  ['o37', 37],
  ['o38', 38],
  ['o39', 39],
  ['o40', 40],
  ['o41', 41],
  ['o42', 42],
  ['o43', 43],
  ['o44', 44],
  ['o45', 45],
  ['o46', 46],
  ['o47', 47],
  ['o49', 49],
  ['o50', 50],
  ['o51', 51],
  ['o52', 52],
  ['o53', 53],
  ['o54', 54],
  ['o55', 55],
  ['o56', 56],
  ['o57', 57],
  ['o58', 58],
  ['o59', 59],
];